Split new nested behavior into different credcache impl
